refactor cube schema location
