+++ tests/unit/backend/wmg/api/test_v1.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 
 from backend.corpora.api_server.app import app
 from backend.wmg.api import v1
-from backend.wmg.data.schema import cube_non_indexed_dims
+from backend.wmg.data.schemas import cube_non_indexed_dims
 from tests.unit.backend.corpora.fixtures.environment_setup import EnvironmentSetup
 from tests.unit.backend.wmg.fixtures.cube import create_temp_cube, all_ones_attr_values
